requirements.txt
ics==0.7.2
requests==2.28.1
bs4==0.0.1
pillow